fix(go): allow go.mod in sub-directories
The current implementation of the go test tooling only works when the
go.mod file is present at the root of the repository.

This commit introduces a `goModPath` option that allow pre-commit nix
configurations to specify the sub-directory that contains the go.mod
file.

All necessary checks use this setting and it will be required when
running this tooling in projects where go.mod lives in a sub-directory.

As part of the refactoring, we also remove implicit shell dependencies
from the test scripts.